# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Parcial 2 - Elecciones 2023'","description":"'Parcial 2 - Elecciones 2023'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1'"},"servers":[{"url":"http://localhost:8081","description":"'Parcial 2 - Elecciones 2023'"}],"paths":{"/secciones":{"get":{"tags":["elecciones-controller"],"operationId":"getSeccionByDistrito","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"seccion_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Seccion"}}}}}}}},"/resultados":{"get":{"tags":["elecciones-controller"],"operationId":"getResultado","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"seccion_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultadoPorDistritoYSeccionDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/distritos":{"get":{"tags":["elecciones-controller"],"operationId":"getDistritoByNombre","parameters":[{"name":"distrito_nombre","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Distrito"}}}}}}}},"/cargos":{"get":{"tags":["elecciones-controller"],"operationId":"getCargoByDistrito","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CargosPorDistritoDTO"}}}}}}}},"components":{"schemas":{"Seccion":{"type":"object","properties":{"seccionId":{"type":"integer","format":"int64"},"seccionNombre":{"type":"string"},"distritoId":{"type":"integer","format":"int64"}}},"ResultadoPorDistritoYSeccionDTO":{"type":"object","properties":{"distrito":{"type":"string"},"seccion":{"type":"string"},"resultados":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoResponseDTO"}}}},"ResultadoResponseDTO":{"type":"object","properties":{"orden":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"votos":{"type":"integer","format":"int64"},"porcentaje":{"type":"number","format":"double"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"Distrito":{"type":"object","properties":{"distritoId":{"type":"integer","format":"int64"},"distritoNombre":{"type":"string"}}},"CargoResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}},"CargosPorDistritoDTO":{"type":"object","properties":{"distrito":{"$ref":"#/components/schemas/DistritoResponseDTO"},"cargos":{"type":"array","items":{"$ref":"#/components/schemas/CargoResponseDTO"}}}},"DistritoResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}}}}}