# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/CargoResponseDTO.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class CargoResponseDTO {

    private Long id;
    private String nombre;

}